src/backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Perchinka <alexsator.lukin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = "^0.24.0.post1"
psycopg2 = "^2.9.9"
passlib = "^1.7.4"
pyjwt = "^2.8.0"
python-multipart = "^0.0.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
